api/requirements.txt
uvicorn>=0.30.1,<=1.0.0
fastapi[standard]>=0.111.0,<1.0.0
pyjwt==2.9.0,<3.0.0
passlib>=1.7.4,<2.0.0
bcrypt==4.1.3
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.1
python-pam==2.0.2
six>=1.16.0
libvirt-python>=11.0.0
psycopg[binary]<4.0.0
psycopg[pool]<4.0.0
streaming-form-data<2.0.0
numpy>=2.3.0,<2.4.0
devtools
cachetools==7.0.5